+++ coveo-example-library/pyproject.toml
@@ -16,7 +16,7 @@ python = ">=3.8"
 coveo-testing = { path = "../coveo-testing", develop = true }
 
 
-[tool.coveo.ci]
+[tool.stew.ci]
 pytest = true
 offline-build = true
 black = true
