+++ coveo-settings/pyproject.toml
@@ -15,7 +15,7 @@ python = ">=3.8"
 coveo-testing = { path = "../coveo-testing", develop = true }
 
 
-[tool.coveo.ci]
+[tool.stew.ci]
 pytest = true
 offline-build = true
 black = true
