+++ coveo-stew/coveo_stew/__main__.py
@@ -2,7 +2,7 @@
 
 if __name__ == "__main__":
     from coveo_systools.script_runner import launch_module_entrypoint_from_sys_argv
-    import coveo_pyproject
+    import coveo_stew
 
     # noinspection PyTypeChecker
-    launch_module_entrypoint_from_sys_argv(coveo_pyproject)
+    launch_module_entrypoint_from_sys_argv(coveo_stew)
