+++ coveo-stew/coveo_stew/exceptions.py
@@ -1,4 +1,4 @@
-"""Exceptions thrown by the pyproject scripts."""
+"""Exceptions thrown by the stew scripts."""
 
 
 class PythonProjectException(Exception):
