+++ coveo-stew/pyproject.toml
@@ -1,14 +1,14 @@
 [tool.poetry]
-name = "coveo-pyproject"
+name = "coveo-stew"
 version = "1.0.0"
-description = "Python packaging and development utilities / poetry on steroids."
+description = "Opinionated python packaging and development utilities"
 license = "Apache-2.0"
 readme = "README.md"
 repository = "https://github.com/coveooss/coveo-python-oss"
 authors = ["Jonathan Piché <tools@coveo.com>"]
 
 packages = [
-    { include = "coveo_pyproject/**/*" },
+    { include = "coveo_stew/**/*" },
 ]
 
 
@@ -37,10 +37,10 @@ pytest = "*"
 
 
 [tool.poetry.scripts]
-pyproject = "coveo_pyproject.commands:pyproject"
+pyproject = "coveo_stew.commands:stew"
+stew = "coveo_stew.commands:stew"
 
-
-[tool.coveo.ci]
+[tool.stew.ci]
 pytest = true
 offline-build = true
 black = true
