+++ coveo-stew/test_coveo_stew/conftest.py
