+++ coveo-stew/test_coveo_stew/pyprojet_mock/mock-pyproject/poetry.toml
