+++ coveo-testing-extras/pyproject.toml
@@ -25,7 +25,7 @@ coveo-testing = { path = "../coveo-testing/", develop = true }
 typing_extensions = "*"
 
 
-[tool.coveo.ci]
+[tool.stew.ci]
 pytest = true
 offline-build = true
 black = true
